refactor(studio): consolidate postJson into api.ts
Move the postJson HTTP utility from http.ts into api.ts and update imports to eliminate the separate http.ts file. This consolidates all Studio API interactions into a single api.ts module.

Also update api.test.ts to properly mock the fetch function instead of the removed http.ts module.
